dashboard/requirements.txt
Flask>=1.0.0
gunicorn==19.7.1
prometheus-client
ibmcloudenv
livereload
mysql-connector-python==8.0.12
flask_wtf
